Move main package from erules.core to erules.
