+++ core/src/main/scala/erules/report/package.scala
@@ -1,4 +1,4 @@
-package erules.core
+package erules
 
 package object report {
   type StringReportEncoder[T] = ReportEncoder[T, String]
