+++ core/src/test/scala/erules/testings/ErulesAsyncAssertingSyntax.scala
@@ -1,7 +1,7 @@
-package erules.core.testings
+package erules.testings
 
 import cats.Functor
-import erules.core.{EngineResult, RuleResult, RuleVerdict}
+import erules.{EngineResult, RuleResult, RuleVerdict}
 import org.scalatest.matchers.should.Matchers
 import org.scalatest.Assertion
