+++ modules/cats-xml/README.md
@@ -25,24 +25,25 @@ case class Person(
 ```
 
 Let's write the rules!
+
 ```scala
-import erules.core.Rule
-import erules.core.PureRule
-import erules.core.RuleVerdict.*
+import erules.Rule
+import erules.PureRule
+import erules.RuleVerdict.*
 import cats.data.NonEmptyList
 import cats.Id
 
 val checkCitizenship: PureRule[Citizenship] =
-  Rule("Check UK citizenship"){
+  Rule("Check UK citizenship") {
     case Citizenship(Country("UK")) => Allow.withoutReasons
-    case _                          => Deny.because("Only UK citizenship is allowed!")
+    case _ => Deny.because("Only UK citizenship is allowed!")
   }
 // checkCitizenship: Rule[Id, Citizenship] = RuleImpl(repl.MdocSession$MdocApp$$Lambda$58763/0x0000000809b592d0@3abe2df3,RuleInfo(Check UK citizenship,None,None))
 
 val checkAdultAge: PureRule[Age] =
-  Rule("Check Age >= 18"){
-    case a: Age if a.value >= 18  => Allow.withoutReasons
-    case _                        => Deny.because("Only >= 18 age are allowed!")
+  Rule("Check Age >= 18") {
+    case a: Age if a.value >= 18 => Allow.withoutReasons
+    case _ => Deny.because("Only >= 18 age are allowed!")
   }
 // checkAdultAge: Rule[Id, Age] = RuleImpl(repl.MdocSession$MdocApp$$Lambda$58767/0x0000000809b782d0@f27dc91,RuleInfo(Check Age >= 18,None,None))
 
@@ -86,7 +87,7 @@ implicit val personEncoder: Encoder[Person] = Encoder.of(person =>
 
 And create the XML report
 ```scala
-import erules.core.*
+import erules.*
 import erules.implicits.*
 import erules.cats.xml.implicits.*
