+++ modules/cats-xml/docs/README.md
@@ -25,23 +25,24 @@ case class Person(
 ```
 
 Let's write the rules!
+
 ```scala mdoc:to-string
-import erules.core.Rule
-import erules.core.PureRule
-import erules.core.RuleVerdict.*
+import erules.Rule
+import erules.PureRule
+import erules.RuleVerdict.*
 import cats.data.NonEmptyList
 import cats.Id
 
 val checkCitizenship: PureRule[Citizenship] =
-  Rule("Check UK citizenship"){
+  Rule("Check UK citizenship") {
     case Citizenship(Country("UK")) => Allow.withoutReasons
-    case _                          => Deny.because("Only UK citizenship is allowed!")
+    case _ => Deny.because("Only UK citizenship is allowed!")
   }
 
 val checkAdultAge: PureRule[Age] =
-  Rule("Check Age >= 18"){
-    case a: Age if a.value >= 18  => Allow.withoutReasons
-    case _                        => Deny.because("Only >= 18 age are allowed!")
+  Rule("Check Age >= 18") {
+    case a: Age if a.value >= 18 => Allow.withoutReasons
+    case _ => Deny.because("Only >= 18 age are allowed!")
   }
 
 val allPersonRules: NonEmptyList[PureRule[Person]] = NonEmptyList.of(
@@ -83,7 +84,7 @@ implicit val personEncoder: Encoder[Person] = Encoder.of(person =>
 
 And create the XML report
 ```scala mdoc:to-string
-import erules.core.*
+import erules.*
 import erules.implicits.*
 import erules.cats.xml.implicits.*
