+++ modules/cats-xml/src/main/scala/erules/cats/xml/report/package.scala
@@ -1,7 +1,7 @@
 package erules.cats.xml
 
 import cats.xml.Xml
-import erules.core.report.ReportEncoder
+import erules.report.ReportEncoder
 
 package object report {
   type XmlReportEncoder[T] = ReportEncoder[T, Xml]
