+++ modules/circe/src/main/scala/erules/circe/report/package.scala
@@ -1,6 +1,6 @@
 package erules.circe
 
-import erules.core.report.ReportEncoder
+import erules.report.ReportEncoder
 import io.circe.Json
 
 package object report {
