+++ modules/generic/src/main/scala-3/erules/generic/RuleMacros.scala
@@ -1,6 +1,6 @@
 package erules.generic
 
-import erules.core.Rule
+import erules.Rule
 
 import scala.annotation.tailrec
