+++ modules/generic/src/test/scala-2/erules/generic/RuleMacrosTest.scala
@@ -1,6 +1,7 @@
 package erules.generic
 
-import erules.core.{PureRule, Rule, RuleVerdict}
+import cats.Id
+import erules.{PureRule, Rule, RuleVerdict}
 import org.scalatest.funsuite.AnyFunSuite
 import org.scalatest.matchers.should.Matchers
