+++ modules/generic/src/test/scala-3/erules/generic/RuleMacrosTest.scala
@@ -1,7 +1,7 @@
 package erules.generic
 
 import cats.Id
-import erules.core.{PureRule, Rule, RuleVerdict}
+import erules.{PureRule, Rule, RuleVerdict}
 import org.scalatest.funsuite.AnyFunSuite
 import org.scalatest.matchers.should.Matchers
